pyproject.toml
[project]
name = "voxdesk"
version = "0.1.0"
description = "VoxDesk — Local-only AI desktop assistant with screen analysis and voice chat"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "ollama>=0.6.0",
    "dxcam>=0.5.0",
    "faster-whisper>=1.1.0",
    "kokoro>=0.9.0",
    "soundfile>=0.13.0",
    "sounddevice>=0.5.0",
    "Pillow>=11.0.0",
    "pyyaml>=6.0",
    "keyboard>=0.13.5",
    "pystray>=0.19.0",
    "numpy>=2.0.0",
    "torch>=2.5.0",
    "websockets>=14.0",
    "av>=14.0",
]

[project.scripts]
voxdesk = "src.main:main"

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-asyncio>=0.25.0",
    "pytest-benchmark>=4.0",
    "pytest-cov>=6.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "unit: Pure logic, no I/O, no hardware",
    "integration: Requires mocked services or real config",
    "regression: Guards against known fixed bugs",
    "benchmark: Performance baseline measurements",
]

[tool.coverage.run]
source = ["src"]
omit = ["src/main.py"]

[tool.coverage.report]
fail_under = 55
show_missing = true
